requirements.txt
aiogram==3.7.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Flask==3.0.0
